+++ ContentPatcher/ModEntry.cs
@@ -79,7 +79,7 @@ internal class ModEntry : Mod
         new Migration_2_4(),
         new Migration_2_5(),
         new Migration_2_6(),
-        new EmptyMigration(2, 7)
+        new Migration_2_7()
     ];
 
     /// <summary>The special validation logic to apply to assets affected by patches.</summary>